+++ tiledesk-client/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Client library for Tiledesk nodejs APIs client
 
-This library is on npm: https://www.npmjs.com/package/@tiledesk/tiledesk-client
+**This library is on npm: https://www.npmjs.com/package/@tiledesk/tiledesk-client**
+
+### 0.8.34
+- Added async/await for getRequestById()
+
+### 0.8.33
+- replaceBotByName() removed hidden start message sending
+
+### 0.8.30
+- added replaceBotByName()
+- fixed findBotByName()
 
 ### 0.8.29
 - added createIntent()